Add u_conn_{f,vf}
The goal is to eventually write a formatter with format specifiers suited
for IRC that can alter output based on connection context. However for now,
we just use vsprintf.

To make this transition easier, *please only use %s and %d (with no parameters)
in your format strings*. An exception is for numerics, where %03d should
be used.