dns: Chain pending requests
i.e., if there's already a pending request for 1.2.3.4.in-addr.arpa,
add our callback + private data pair to a list of interested parties
when adding new requests for that record.

Also, the latter half of u_dns and u_rdns has been replaced with a
single function "request" that takes a record type and name and does the
actual request, checking for cached results and chaining opportunities
as necessary.