Add streq to reduce confusion of !strcmp
I have twice now made the mistake of replacing !strcmp() with !match() or
similiar. The return value of a match()-family function has the opposite
sense of strcmp()-family one. match() functions return 0 to indicate
*no match*, while strcmp() does the opposite. To prevent mistakes of
this nature, and to make it easier to read code using strcmp, the streq
macro was created.